refactor SnapshotUtility.kt into /database
